+++ tests/unit/test_citations.py
@@ -124,14 +124,14 @@ class TestCitations(unittest.TestCase):
         citations._reset()
         self.assertNotIn("Subramanian2005", citations._papers_to_cite)
         pybamm.particle.no_distribution.XAveragedPolynomialProfile(
-            None, "Negative", "quadratic profile"
+            None, "Negative", "quadratic profile", None
         )
         self.assertIn("Subramanian2005", citations._papers_to_cite)
 
         citations._reset()
         self.assertNotIn("Subramanian2005", citations._papers_to_cite)
         pybamm.particle.no_distribution.PolynomialProfile(
-            None, "Negative", "quadratic profile"
+            None, "Negative", "quadratic profile", None
         )
         self.assertIn("Subramanian2005", citations._papers_to_cite)